+++ tasks/pick_path.rb
@@ -1,0 +1,28 @@
+#!/usr/bin/env ruby
+
+task_helper = [
+  # During a real bolt call, ruby_task_helper modules is installed in same directory as this module
+  File.join(__dir__, '..', '..', 'ruby_task_helper', 'files', 'task_helper.rb'),
+  # During development the ruby_task_helper module will be in the test module fixtures
+  File.join(__dir__, '..', 'spec', 'fixtures', 'modules', 'ruby_task_helper', 'files', 'task_helper.rb'),
+].find { |helper_path| File.exist?(helper_path) }
+raise 'Could not find the Bolt ruby_task_helper' if task_helper.nil?
+require_relative task_helper
+
+# Task to find the first file path that exists
+class PickPathTask < TaskHelper
+  def task(paths: nil, **_kwargs)
+    # expand_path to resolve things like ~/ and ../
+    # check for file existence
+    result = paths.find { |p| File.exist?(File.expand_path(p)) }
+    return { value: result }
+  rescue TaskHelper::Error => e
+    # ruby_task_helper doesn't print errors under the _error key, so we have to
+    # handle that ourselves
+    return { _error: e.to_h }
+  end
+end
+
+if $PROGRAM_NAME == __FILE__
+  PickPathTask.run
+end